Worker for loading and composing GeoTIFF tiles
